refactor(cloud-storage): adjust file path
