+++ src/v1/controller/cloudStorage/alibabaCloud/upload/Utils.ts
@@ -1,5 +1,5 @@
-import { CloudStorageConfigsDAO } from "../../../dao";
-import { CloudStorage } from "../../../../Constants";
+import { CloudStorageConfigsDAO } from "../../../../dao";
+import { CloudStorage } from "../../../../../Constants";
 
 export const checkTotalUsage = async (
     userUUID: string,
